Add request:log() function
This allows some basic logging to stdio. Takes a level string (WARN, ERROR, INFO) and the message to print, and gives a somewhat useful output. I'm not sure I'm happy with the format, it should probably conform to NGINX or apache access logs style. Logging to a file vs blindly printing may also be useful.